+++ mxml.h
@@ -141,7 +141,7 @@ extern void		mxmlDelete(mxml_node_t *node);
 
 extern void		mxmlElementClearAttr(mxml_node_t *node, const char *name);
 extern const char	*mxmlElementGetAttr(mxml_node_t *node, const char *name);
-extern const char       *mxmlElementGetAttrByIndex(mxml_node_t *node, int idx, const char **name);
+extern const char       *mxmlElementGetAttrByIndex(mxml_node_t *node, size_t idx, const char **name);
 extern size_t		mxmlElementGetAttrCount(mxml_node_t *node);
 extern void		mxmlElementSetAttr(mxml_node_t *node, const char *name, const char *value);
 extern void		mxmlElementSetAttrf(mxml_node_t *node, const char *name, const char *format, ...) MXML_FORMAT(3,4);
